refactor: Added abstraction for NotificationDataSource
